requirements.txt
textual>=0.80.0
textual-plotext>=1.0.0
plotext>=5.3.0
anthropic>=0.40.0
pyyaml>=6.0
